Move metricstorage consts to api
