+++ api/v1beta1/metricstorage_consts.go
@@ -13,7 +13,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package metricstorage
+package v1beta1
 
 import "time"
